#100 #101 Fixes issues with tool feedback disposal
- Fix incorrect handling of EnableDefaultTools action in tool palette
- Ensure that the correct FeedbackEmitters are unregistered on tool deactivation.
- Fix a bug  that was sending empty CompoundActions to the server when panning the canvas
- Make EdgeEditTool inherit from BaseGLSPTool and avoid construcutor injection
- Ensure that no unnecessary FeedbackActions are dispatched (e.g. showing resizehandles for non-resizable elements)
- Split `creation-tool.ts` into two separate files
- Move dragaware-mouse-listener into base directory.
- Organize index imports
- Ensure that DeleteKeyToolListener does not dispatch an action on empty selection

Fixes https://github.com/eclipse-glsp/glsp/issues/100
Fixes https://github.com/eclipse-glsp/glsp/issues/101

+++ src/features/tools/delete-tool.ts
@@ -62,7 +62,9 @@ export class DeleteKeyListener extends KeyListener {
         if (matchesKeystroke(event, 'Delete')) {
             const deleteElementIds = Array.from(element.root.index.all().filter(e => isDeletable(e) && isSelectable(e) && e.selected)
                 .filter(e => e.id !== e.root.id).map(e => e.id));
-            return [new DeleteElementOperation(deleteElementIds)];
+            if (deleteElementIds.length > 0) {
+                return [new DeleteElementOperation(deleteElementIds)];
+            }
         }
         return [];
     }
@@ -99,7 +101,6 @@ export class DeleteToolMouseListener extends MouseListener {
         if (!isDeletable(target)) {
             return [];
         }
-
         const result: Action[] = [];
         result.push(new DeleteElementOperation([target.id]));
         if (!isCtrlOrCmd(event)) {
